% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cells}
\alias{cells}
\title{Simulated single cell values (phenotypes).}
\format{A list with 1490 item, each item is a vector of 20000 integer}
\source{
\url{http://membres-timc.imag.fr/Daniel.Jost/}
}
\usage{
cells
}
\description{
A list of the 20000 single cell values (phenotype) of the 1490 individuals described by the variable 'indiv'.
}
\keyword{datasets}

