% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genodata.R
\name{compute_teta}
\alias{compute_teta}
\title{A Function That Computes Recombination Factors.}
\usage{
compute_teta(genodata)
}
\arguments{
\item{genodata}{A genome generated with the function `generate_genodata`}
}
\description{
This function computes recombination factors between markers.
}

