% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genodata}
\alias{genodata}
\title{Genotype of the simulated individuals.}
\format{A matrix with 210 rows and 1490 columns}
\source{
\url{http://membres-timc.imag.fr/Daniel.Jost/}
}
\usage{
genodata
}
\description{
A matrix whose rows are markers and columns are individuals. The marker 101 corresponds to the 'all' variable of the 'indiv' data frame, it is the parental allele linked with the phenotype. From this marker 101, genotype is propagated (upstream and downstream) with recombination factor of 5%.
}
\keyword{datasets}

