% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiv}
\alias{indiv}
\title{Description of the simulated individuals.}
\format{A data frame with 1490 rows and 8 variables:
\describe{
  \item{err}{The value of the error introducted in the paramater selection level.}
  \item{all}{The parental allele.}
  \item{a0}{The a0 parameter value.}
  \item{a1}{The a1 parameter value.}
  \item{k}{The k parameter value.}
  \item{m1}{The first moment of the single cell distribution.}
  \item{m2}{The second moment of the single cell distribution.}
  \item{m3}{The third moment of the single cell distribution.}
}}
\source{
\url{http://membres-timc.imag.fr/Daniel.Jost/}
}
\usage{
indiv
}
\description{
A data frame that contains the description of the 1490 individuals.
}
\keyword{datasets}

